Move exceptions handling to a separate module
